+++ scripts/macros/feats/maelstromAura.js
@@ -1,0 +1,26 @@
+import {constants} from '../../constants.js';
+import {chris} from '../../helperFunctions.js';
+import {queue} from '../../utility/queue.js';
+export async function maelstromAura(token, origin) {
+    console.log('here');
+    let targetToken = game.canvas.tokens.get(game.combat.current.tokenId);
+    if (!targetToken) return;
+    console.log('here');
+    if (targetToken.document.disposition === token.document.disposition) return;
+    console.log('here');
+    let distance = chris.getDistance(token, targetToken);
+    if (distance > 10) return;
+    console.log('here');
+    let featureData = await chris.getItemFromCompendium('chris-premades.CPR Feat Features', 'Maelstrom Aura', false);
+    if (!featureData) return;
+    console.log('here');
+    featureData.system.save.dc = chris.getSpellDC(origin);
+    featureData.system.description.value = chris.getItemDescription('CPR - Descriptions', 'Maelstrom Aura');
+    let feature = new CONFIG.Item.documentClass(featureData, {'parent': token.actor});
+    let [config, options] = constants.syntheticItemWorkflowOptions([targetToken.document.uuid]);
+    let queueSetup = await queue.setup(origin.uuid, 'maelstromAura', 50);
+    if (!queueSetup) return;
+    await warpgate.wait(100);
+    await MidiQOL.completeItemUse(feature, config, options);
+    queue.remove(origin.uuid);
+}